feature(listar_contratos): adicionando listagem de contratos no menu lateral
