+++ src/pages/contract/components/styles.scss
@@ -1,0 +1,98 @@
+.sidebar {
+  width: 70px;
+  background-color: #cfcfcf;
+  transition: width 0.3s ease;
+  height: 100%;
+  padding: 10px;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+
+  &:hover {
+    width: 280px;
+    overflow-y: auto;
+
+    .top-bar {
+      .field-search,
+      .btn-icon {
+        display: inline-block;
+      }
+    }
+
+    .user .username {
+      display: inline;
+    }
+  }
+
+  .top-bar {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 12px;
+    margin-bottom: 5px;
+
+    .field-search,
+    .btn-icon {
+      display: none;
+    }
+
+    .search-row {
+      display: flex;
+      align-items: center;
+      flex: 1;
+    }
+
+    .field-search {
+      flex: 1;
+      padding: 10px 10px;
+      border-radius: 10px;
+      border: none;
+    }
+
+    .btn-icon {
+      background: none;
+      border: none;
+      color: #000080;
+      padding: 10px;
+      font-size: 24px;
+      cursor: pointer;
+    }
+  }
+
+  .user {
+    display: flex;
+    align-items: center;
+    padding: 8px 5px;
+    gap: 10px;
+    cursor: pointer;
+
+    &:hover {
+      background-color: #e0e0e0;
+    }
+
+    .icon-user {
+      width: 40px;
+      height: 40px;
+      border-radius: 50%;
+      background-color: #f1f1f1;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      font-size: 24px;
+      margin: 10px 0;
+      flex-shrink: 0;
+    }
+
+    i.bi-person-fill {
+      font-size: 24px;
+      color: #000080;
+    }
+
+    .username {
+      display: none;
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
+  }
+}
